vortex: Fixup unembed layer usage when not using tie_embeddings
Last commit didn't account for tied_embeddings=False case and used
weights from `embed` layer unconditionally.

This commit fixes it.

Are some nuances to be aware of:

When not using tie_embedding, we must still assign new layer into
`unembed` member variable (as this is where weights are expected to be
loaded), but then if we want to keep layer name the same for the purpose
of installing the hook, we cannot use the Lambda wrapper, we must
instantiate proper class.

(cherry picked from commit 80d9df0a3c75abdf27a3815d771ab06c326c7ba9)